refactor: separate survey form component
